Thread-based processing
In non-threaded approach on host with low computing power, sometimes
the overrun has occurred (in the ALSA internal buffer). It was due to
the fact, that the audio buffer was not drained fast enough, because
the reading and writing routines were not separated. The writing routine
might have temporal requirement for high computation power or high
disk output latency.