Do not copy PCM data between local buffers
Instead of keeping PCM data in a local buffer in every thread, use
global ring buffer which will store captured PCM. The access to the
buffer itself is shared between threads, but access to ring buffer
pointers are protected with a mutex.

Profiling done with callgrind shows ~3.5% performance improvement
when used with WAV writer.